requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
openai>=1.0.0
Pillow>=9.5.0
python-multipart>=0.0.6
jinja2>=3.1.2
itsdangerous>=2.1.2
aiofiles>=23.2.1
